src/requirements.txt
prefect==2.20.3
openpyxl==3.1.2
python-gitlab==5.0.0
pydantic==2.11.0
pydantic_yaml==1.2.1
lz4==4.3.2
numpy==1.26.2
pandas==2.1.3
unidecode==1.3.8
docker==6.1.3
junit-xml==1.9
psycopg2-binary==2.9.9
keycove==0.3.5
fastapi==0.109.0
sqlalchemy_utils==0.41.2
click==8.1.7
alive-progress==3.2.0
gitpython==3.1.43
aiocache[redis]==0.12.3
SQLAlchemy==1.4.54
alembic==1.15.1
Faker==37.1.0
dotenv==0.9.9